1- Changed WIT ID type from Double to Int 2- Added toolbar for login information 3- Copied information to clipboard 4- Displaying the HTML version (even if the font is crooked)

+++ Formatho/ADO.swift
@@ -129,12 +129,12 @@ class ADOWitSearch:Codable, Identifiable {
 class Wit: Codable, Identifiable {
     
     init() {
-        id = Double()
+        id = Int()
         fields = Fields()
         url = String()
     }
     
-    let id: Double
+    let id: Int
     let fields: Fields
     let url: String
 }
